Se coloco mensaje de confirmacion en la peticion delete

+++ js/peticionesDelete.js
@@ -1,19 +1,22 @@
 const eliminarTarea = ( id_tarea ) => {
 
-    let _datos = {
-        tarea: id_tarea
-    }
+    if ( confirm("¿Desea eliminar la tarea?") == true) {
+        let _datos = {
+            tarea: id_tarea
+        }
 
-    fetch('http://proyectoeasy.shop/backend_todo/tarea/tarea.php', {
-        method: "DELETE",
-        body: JSON.stringify(_datos),
-        headers: { "Content-type": "application/json; charset=UTF-8" }
-    })
-        .then(response => response.json())
-        .then(json => {
-            if (json.results?.id) {
-                location.reload();
-            }
+        fetch('http://proyectoeasy.shop/backend_todo/tarea/tarea.php', {
+            method: "DELETE",
+            body: JSON.stringify(_datos),
+            headers: { "Content-type": "application/json; charset=UTF-8" }
         })
-        .catch(err => console.log(err));
+            .then(response => response.json())
+            .then(json => {
+                if (json.results?.id) {
+                    location.reload();
+                }
+            })
+            .catch(err => console.log(err));
+
+    }
 }
